Add bundle build endpoint
Refactor to move the handlers under server
Remove afero where it's not needed
Bundle build endpoint accepts a POST request and will return a 200 or error
